WEB-272 replace custom api of home.service.ts
