+++ src/app/customApis.service.ts
@@ -231,3 +231,65 @@ export class OrganizationService {
     });
   }
 }
+
+/**
+ * Home Service
+ */
+@Injectable({
+  providedIn: 'root'
+})
+export class HomeService {
+  /**
+   * @param {HttpClient} http Http Client to send requests.
+   */
+  constructor(private http: HttpClient) {}
+
+  /**
+   * @param {number} officeId Office Id.
+   * @returns {Observable<any>}
+   */
+  getCollectedAmount(officeId: number): Observable<any> {
+    const httpParams = new HttpParams().set('R_officeId', officeId.toString()).set('genericResultSet', 'false');
+    return this.http.get('/runreports/Demand Vs Collection', { params: httpParams });
+  }
+
+  getDisbursedAmount(officeId: number): Observable<any> {
+    const httpParams = new HttpParams().set('R_officeId', officeId.toString()).set('genericResultSet', 'false');
+    return this.http.get('/runreports/Disbursal Vs Awaitingdisbursal', { params: httpParams });
+  }
+
+  /**
+   * @param {number} officeId Office Id.
+   * @returns {Observable<any>}
+   */
+
+  getClientTrendsByDay(officeId: number): Observable<any> {
+    const httpParams = new HttpParams().set('R_officeId', officeId.toString()).set('genericResultSet', 'false');
+    return this.http.get('/runreports/ClientTrendsByDay', { params: httpParams });
+  }
+
+  getLoanTrendsByDay(officeId: number): Observable<any> {
+    const httpParams = new HttpParams().set('R_officeId', officeId.toString()).set('genericResultSet', 'false');
+    return this.http.get('/runreports/LoanTrendsByDay', { params: httpParams });
+  }
+
+  getClientTrendsByWeek(officeId: number): Observable<any> {
+    const httpParams = new HttpParams().set('R_officeId', officeId.toString()).set('genericResultSet', 'false');
+    return this.http.get('/runreports/ClientTrendsByWeek', { params: httpParams });
+  }
+
+  getClientTrendsByMonth(officeId: number): Observable<any> {
+    const httpParams = new HttpParams().set('R_officeId', officeId.toString()).set('genericResultSet', 'false');
+    return this.http.get('/runreports/ClientTrendsByMonth', { params: httpParams });
+  }
+
+  getLoanTrendsByWeek(officeId: number): Observable<any> {
+    const httpParams = new HttpParams().set('R_officeId', officeId.toString()).set('genericResultSet', 'false');
+    return this.http.get('/runreports/LoanTrendsByWeek', { params: httpParams });
+  }
+
+  getLoanTrendsByMonth(officeId: number): Observable<any> {
+    const httpParams = new HttpParams().set('R_officeId', officeId.toString()).set('genericResultSet', 'false');
+    return this.http.get('/runreports/LoanTrendsByMonth', { params: httpParams });
+  }
+}
